pyproject.toml
[project]
name = "ha-services"
dynamic = ["version"]
description = "Helpers to send periodic information via MQTT to Home Assistant"
license = {text = "GPL-3.0-or-later"}
readme = "README.md"
authors = [
    {name = 'Jens Diemer', email = 'github@jensdiemer.de'}
]
requires-python = ">=3.11"
dependencies = [
    "paho-mqtt",  # https://pypi.org/project/paho-mqtt/
    "tomlkit",  # https://github.com/sdispater/tomlkit
    "msgspec",  # https://github.com/jcrist/msgspec
    "frozendict",  # https://github.com/Marco-Sulla/python-frozendict
    "psutil",  # https://github.com/giampaolo/psutil/
    "cli-base-utilities>=0.18.0",  # https://github.com/jedie/cli-base-utilities
    "tyro",  # https://github.com/brentyi/tyro
    "rich",  # https://github.com/Textualize/rich
]

[dependency-groups]
dev = [
    "freezegun",  # https://github.com/spulec/freezegun
    "manageprojects",  # https://github.com/jedie/manageprojects
    "urllib3", # for bx_py_utils.test_utils.deny_requests.deny_any_real_request() in tests
    "uv",  # https://github.com/astral-sh/uv
    "setuptools",  # https://github.com/pypa/setuptools
    "nox",  # https://github.com/wntrblm/nox
    "coverage",  # https://github.com/nedbat/coveragepy
    "autopep8",  # https://github.com/hhatto/autopep8
    "pyupgrade",  # https://github.com/asottile/pyupgrade
    "flake8",  # https://github.com/pycqa/flake8
    "flake8-bugbear",  # https://github.com/PyCQA/flake8-bugbear
    "pyflakes",  # https://github.com/PyCQA/pyflakes
    "codespell",  # https://github.com/codespell-project/codespell
    "EditorConfig",  # https://github.com/editorconfig/editorconfig-core-py
    "pip-audit",  # https://github.com/pypa/pip-audit
    "mypy",  # https://github.com/python/mypy
    "twine",  # https://github.com/pypa/twine
    "pre-commit",  # https://github.com/pre-commit/pre-commit
    "typeguard",  # https://github.com/agronholm/typeguard/

    # https://github.com/akaihola/darker
    # https://github.com/ikamensh/flynt
    # https://github.com/pycqa/isort
    # https://github.com/pygments/pygments
    "darker[flynt, isort, color]",
]

[project.urls]
Documentation = "https://github.com/jedie/ha_services"
Source = "https://github.com/jedie/ha_services"

[project.scripts]
ha_services_app = "ha_services.__main__:main"
ha_services_dev = "ha_services.cli_dev:main"

[build-system]
requires = ["setuptools>=61.0", "setuptools_scm>=7.1"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["ha_services*"]

[tool.setuptools.dynamic]
version = {attr = "ha_services.__version__"}


[tool.cli_base.pip_audit]
requirements=["requirements.dev.txt"]
strict=true
require_hashes=true
ignore-vuln=[]


[tool.cli_base]
version_module_name = "ha_services"  # Used by "update-readme-history" git hook


[tool.darker]
src = ['.']
revision = "origin/main..."
line_length = 119
color = true
skip_string_normalization = true
diff = false
check = false
stdout = false
isort = true
log_level = "INFO"


[tool.isort]
# https://pycqa.github.io/isort/docs/configuration/config_files/#pyprojecttoml-preferred-format
atomic=true
profile='black'
skip_glob=['.*', '*/htmlcov/*']
known_first_party=['ha_services']
line_length=119
lines_after_imports=2


[tool.coverage.run]
branch = true
parallel = true
concurrency = ["multiprocessing"]
source = ['.']
command_line = '-m unittest --verbose --locals --buffer'

[tool.coverage.report]
omit = ['.*', '*/tests/*']
skip_empty = true
fail_under = 30
show_missing = true
exclude_lines = [
    'if self.debug:',
    'pragma: no cover',
    'raise NotImplementedError',
    'if __name__ == .__main__.:',
]


[tool.mypy]
warn_unused_configs = true
ignore_missing_imports = true
allow_redefinition = true  # https://github.com/python/mypy/issues/7165
show_error_codes = true
plugins = []
exclude = ['.venv', 'tests']


[manageprojects] # https://github.com/jedie/manageprojects
initial_revision = "a264778"
initial_date = 2023-05-07T13:44:43+02:00
cookiecutter_template = "https://github.com/jedie/cookiecutter_templates/"
cookiecutter_directory = "uv-python"
applied_migrations = [
    "765b8fb", # 2024-03-12T09:08:01+01:00
    "935485b", # 2024-09-09T10:39:39+02:00
    "b7f2cb5", # 2025-03-23T15:08:42+01:00
]

[manageprojects.cookiecutter_context.cookiecutter]
full_name = "Jens Diemer"
github_username = "jedie"
author_email = "github@jensdiemer.de"
package_name = "ha_services"
package_version = "2.8.0"
package_description = "Helpers to send periodic information via MQTT to Home Assistant"
package_url = "https://github.com/jedie/ha_services"
issues_url = "https://github.com/jedie/ha_services/issues"
license = "GPL-3.0-or-later"
_template = "https://github.com/jedie/cookiecutter_templates/"
applied_migrations = [
    "363c360", # 2023-12-16T19:37:39+01:00
]
